refactor: add common folder and move related files in it
