+++ androidApp/src/main/java/com/example/tasky/android/login/viewmodel/LoginViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.android.login.screen.LoginScreenEvent
 import com.example.tasky.android.login.screen.LoginScreenState
+import com.example.tasky.common.util.Validator
 import com.example.tasky.login.data.model.LoginBody
 import com.example.tasky.login.domain.loginManager
-import com.example.tasky.util.Validator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
