+++ androidApp/src/main/java/com/example/tasky/android/login/viewmodel/RegisterViewModel.kt
@@ -4,11 +4,11 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.android.login.screen.RegisterScreenEvent
 import com.example.tasky.android.login.screen.RegisterScreenState
+import com.example.tasky.common.model.onError
+import com.example.tasky.common.model.onSuccess
+import com.example.tasky.common.util.Validator
 import com.example.tasky.login.data.model.RegisterBody
 import com.example.tasky.login.domain.ILoginRepository
-import com.example.tasky.model.onError
-import com.example.tasky.model.onSuccess
-import com.example.tasky.util.Validator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
