+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/domain/AgendaRepository.kt
@@ -10,8 +10,8 @@ import com.example.tasky.agenda.data.model.GetAgendaResponse
 import com.example.tasky.agenda.data.model.Reminder
 import com.example.tasky.agenda.data.model.Task
 import com.example.tasky.agenda.data.model.UpdateEventBody
-import com.example.tasky.model.BaseError
-import com.example.tasky.model.ResultWrapper
+import com.example.tasky.common.model.BaseError
+import com.example.tasky.common.model.ResultWrapper
 
 interface IAgendaRepository {
     suspend fun getAgenda(timeStamp: Long): ResultWrapper<GetAgendaResponse, BaseError>
