+++ shared/src/commonMain/kotlin/com/example/tasky/common/manager/HttpManager.kt
@@ -1,11 +1,11 @@
-package com.example.tasky.manager
+package com.example.tasky.common.manager
 
 import com.example.tasky.BuildKonfig.API_KEY
+import com.example.tasky.common.model.ErrorResponse
+import com.example.tasky.common.util.isSuccess
 import com.example.tasky.login.data.model.AccessToken
 import com.example.tasky.login.data.model.AccessTokenResponse
 import com.example.tasky.login.domain.SessionManager
-import com.example.tasky.model.ErrorResponse
-import com.example.tasky.util.isSuccess
 import com.russhwolf.settings.ExperimentalSettingsApi
 import com.russhwolf.settings.ExperimentalSettingsImplementation
 import io.ktor.client.HttpClient
