+++ shared/src/commonMain/kotlin/com/example/tasky/common/model/BaseError.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.model
+package com.example.tasky.common.model
 
 interface BaseError
