+++ shared/src/commonMain/kotlin/com/example/tasky/common/model/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.model
+package com.example.tasky.common.model
 
 import kotlinx.serialization.Serializable
