+++ shared/src/commonMain/kotlin/com/example/tasky/login/data/LoginDataSource.kt
@@ -1,14 +1,14 @@
 package com.example.tasky.login.data
 
+import com.example.tasky.common.manager.HttpManager
+import com.example.tasky.common.model.BaseError
+import com.example.tasky.common.model.ResultWrapper
+import com.example.tasky.common.util.safeCall
 import com.example.tasky.login.data.model.Login
 import com.example.tasky.login.data.model.LoginBody
 import com.example.tasky.login.data.model.LoginResponse
 import com.example.tasky.login.data.model.Register
 import com.example.tasky.login.data.model.RegisterBody
-import com.example.tasky.manager.HttpManager
-import com.example.tasky.model.BaseError
-import com.example.tasky.model.ResultWrapper
-import com.example.tasky.util.safeCall
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.resources.post
 import io.ktor.client.request.setBody
