+++ shared/src/commonMain/kotlin/com/example/tasky/login/domain/LoginRepository.kt
@@ -1,11 +1,11 @@
 package com.example.tasky.login.domain
 
+import com.example.tasky.common.model.BaseError
+import com.example.tasky.common.model.ResultWrapper
 import com.example.tasky.login.data.LoginDataSource
 import com.example.tasky.login.data.model.LoginBody
 import com.example.tasky.login.data.model.LoginResponse
 import com.example.tasky.login.data.model.RegisterBody
-import com.example.tasky.model.BaseError
-import com.example.tasky.model.ResultWrapper
 
 interface ILoginRepository {
     suspend fun login(loginBody: LoginBody): ResultWrapper<LoginResponse, BaseError>
